DELIA-30990 : Destroying pipeline without changing it to NULL
Reason for change: Changes to avoid setting the pipeline to NULL was brought as part
of change 19692 (DELIA-27445).
In the case of "video decode error" pipeline STATE_CHANGE is in FAILURE state, but no hang
is observed if we try to set pipeline to NULL.
Tried to reproduce the crash when pipeline was in failure state via CATR-7825.
But crash was not reproduced even when pipeline was in failure state.
This change modifies change 19692; To go ahead with setting pipeline to NULL
even if it's in failure state, and adds prints to monitor if pipeline is in
failed state.

Test Procedure: Channel change test in rack and see hang/crash is seen.
Risks: Medium
Signed-off-by: Akhil Babu <Akhil_Babu2@comcast.com>

Change-Id: I6b519112f808c77f346180b9c90c0d1e9080c285