DELIA-31993 Crash when do channel up for America's Newsroom.
Reason for change: This crash is specifically seen with FNCHD
channel. Crash was due to no segmentUrls on the dash manifest.
Added check point to avoid the crash when there is no segmentUrls.
Also added precaution checks where ever the segmentUrls are accessed.
Test Procedure: Tested by simulation. Need to verify this in Live.
Risks: None
Signed-off-by: Paulpandian M paulpandian_mariappan@cable.comcast.com

Change-Id: If6ade62a6d14c3c9713a8532ceab47448cd055d3