NDVR-4386 Fail-over - Stream switching to secondary taking long time
Reason for change: To handle an edge case for cdn failover test, where
we found valid sub-manifest but no valid fragments. In this case
we have to stall the playback. So that the retry handler will try
for next available server location if multi location
is available for the asset.
Test Procedure: Tested by simulation on Xi5
Risks: Low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: I571c8478a18c23ff1d3f814c3925397e2f002907