DELIA-34286 FIELD ROGERS CDVR Increase in AAMP: init failures with 3.4p5s1
Reason for change: Curl download failures/timeouts for audio video playlist
are not retried since the buffer->len is 0 on that time (and this is an edge
case observed on rogers environment). Modified the condition to retry audio
video playlist on timeouts (or) if the buffer is partial on timeout.
Test Procedure: Tested with Xi5
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Iaa23ad5f023c15e5ffab4c84aeb10e0552b13d8a