DELIA-34308 [Field] [Data Analysis] Webprocess crash at TrackState::Stop
Reason for change: At an edge case the SetRate is called after
a playback failed due to manifest download failure at seek point 0
*** At that time, pipeline was already NULL, and the Pause()
operation was ignored but the protection mutex which is locked
before this operation is not release that is the reason for
the hung come crash.
Along with this, added below patches to improve this case.
* Avoided the call of NotifySpeedChanged() is Pause/SetRate is failed.
* Resolved PC simulator compilation error on fragmentcollector_hls.h
with timeval st.
Test Procedure: Tested by simulation on Xi6.
Risks: Low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Ie177f8511b04c775bcd420ede7024f7aae12f73d