DELIA-34803 [EAS] [field] Amber Alert started mid-way through alert
Reason for change: Current logic is not considering EAS as VOD.
So the liveAdjust offset has been calculated as like for Live and
applied for this EAS playback. Now this has been handled on IsLiveAdjustRequired().
Test Procedure: Unable to verify with current live stream since its an EAS,
so simulated this case and verified the patch.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: I1b32afde976310e574eba545f3996154b6bc1ded