DELIA-34039: stall detection for CURL download requests
Reason for change: Remove low speed configuration based stall detection
logic and replace with a logic in progress callback on which we have more
control. Two new configs curl-stall-timeout and curl-download-start-timeout
has been brought in to to configure stall timeouts for mid-download and
no-data freeze scenarios.
Test Procedure: tested on arris xg1v4 with FOG/DAI assets to check for regression
and in PC with a custom server which can reproduce both types of stall.
Risks: Low
Signed-off-by: Vinish K B <Vinish_KorasseryBalakrishnan@comcast.com>

Change-Id: I7e03c02ce2c6d0e3f6e749fec025aaf86674c2d5