RDK-24188 : new onVideoEnd event to report statistics
Reason for change: added Support for Collecting video stats. 
and Generating Metric Event on next tune attmept. Trace ID from XRE will be used for 
mapping with tune request. 
with this implemention license stats and Fog integration is pending. 

Test Procedure: Tested on XI5 with info prints. Event is not reaching to XRE as XRE needs change to register or listen to event. 

Change-Id: I699172344dff1e397804dd07d4ea2b7a87e673cb
Risks:Low
Signed-off-by: shripad bankar <Shripad_Bankar@cable.comcast.com>