DELIA-35927: Audio issue when FF/REW multiple times in succession
Reason for change: mFirstPTS is not set in this scenario which
causes audio fragments to be out of sync resulting in audio drop.
It appears to be related to this particular asset and the seek position.
Test Procedure: tested on arris xg1v4
Risks: None
Signed-off-by: Vinish K B <Vinish_KorasseryBalakrishnan@comcast.com>

Change-Id: I7153f0a69e084fb785657e689920eb7f54a74f78
(cherry picked from commit 398657aa010bf488d04535a9d9ccbbd561f03ffd)