DELIA-35707 [AAMP][IPVOD][Intermittent] Blackscreen at end of IPVOD(DAI enabled) playback
Reason for change: Audio duration is greater than video duration just before end of playback,
in  that scenario, audio is going to be wait to catch up video track to match duration, but for that time
video injection is completed and there is no more video track to catch up; in that case
audio is still in wait to catch up video track and there is no body to signal it.
Fixed the issue by avoiding the call of pthread_cond_wait() for audio injector thread
if video injection was already exited/done.
Test Procedure: Tested in Xi5 with Superstore S4 Ep15.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: I8440d3348b35f8aac0d155b909bd929c30846174