DELIA-36930 misleading AAMP ABR logging
Reason for change: On some streams profiles are organized highest to lowest,
and so each ramp down is incorrectly logged as ramping "higher"
and wise-versa for ramp up.
Changed the condition with respect to bandwidth instead of profile index
to rectify the incorrect logging.
Test Procedure: Tested in Xi5
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Ib09b7a317d98bf628e1addc5841df0d3fbea9366