BCOM-3558 : Pipeline error due to session manager inactive
Reason for change:
Added changes to return the drm session, if a session is already
cached for the requested keyId. Moved the in-active check (which was
introduced to avoid aamp request license for failed keyId, right after
a stop call) further below to acheive this.

Test Procedure: Test DASH drm, success and failure case (auth failed)
Risks: Low
Signed-off-by: Akhil Babu <Akhil_Babu2@comcast.com>

Change-Id: I8d68ac82d04eec370006402916d2d7ef5e22f9d6