DELIA-41062 [AAMP] [PEACOCK] Freeze at AD transition in Linear
Reason for change: Audio discontinuity was ignored in a random case
where the video playlist updates faster, then after the video discontinuity
updates and that is getting processed, on that time
there is no audio discontinuity available to process.
This is observed in Peacock linear case.
Test Procedure: Tested with Xi5
Risks: Low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Ifd0c9e45a4794f8899529a6c64aa8ff54415914d