DELIA-41375 [AAMP] [PEACOCK] Video scaling issue when enabled Westeros for Peacock.
Reason for change: There are no AAMPWebVideoItems
in the case of the Media player to handle the scaling,
so Player level scaling is handled to fix
the issue even Westeros is enabled.
Test Procedure: Tested play backing XRE playback and
Peacock playbacks and did not see any scaling issues.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: I6a4aebef343a5db0b62250735aabc32945008851