DELIA-41666 [Peacock] [VOD] Temporary audio loss when rewind seek initiated
Reason for change: Due to indexCount mismatch calculating the individual
play target has ~1 to 2 sec diff comparing either of tracks and that
makes the issue here, but in this case, the period start position is same
for audio and video. So the change added to find the play target
for both audio and video based on video's offsetFromPeriod
if the period start position same for both the tracks.
Test Procedure: Tested play backing American Ninja asset
through Peacock.ref app and seeking to position 3800,
audio is played fine. No drop has been seen.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: I6c11f882cc0d8fd9259ecd43c28910622bd9af9c