DELIA-39828: Crash at EOS Event() processing.
Reason for change: During EOS event processing App blocks the event call
and issued a release for event listeners, during the release event
listeners the object was deleted and the pointe became NULL,
so with that NULL calling JSValueUnprotect crashed.
Test Procedure: Seek the playback to the end in Peacock ref App then leave it,
there should not be any crash at the playback end.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: I57b8007a746e4e99654db9e79a5cf57b0d30e8fc