XIONE-373 RDK-25745 (VCG custom build) DASH PlayReady/WV playback failing with 'Failed to locate DRM helper'
Reason for change: Drm preference was not considered for WiveVine on mDrmPrefs,
so per the default preference, Playready was considered as preferred DRM and
the key retrieved for PlayReady then for the WideVine got processed,
in that state, the key cached was having the PlayReady which was not matched
with WideVine then a new slot was created and this license request for this WideVine failed. Which is not a usual case.
Fixed by giving the preference for the DRM based on GetPreferredDRM() call.
Test Procedure: Tested by configuring preferred-drm=1 (or) preferred-drm=2, now the playback goes fine.
Risks: Low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Ifc0009798b92a32c71933b95748ba6bf92a95556