DELIA-40680:Page up towards end of IP-VOD or CDVR asset does not exit
Reason for change: To send an EOS when the playtarget is
greater than or equal to the seekwindow rounded to avoid
floating point problem.
Test Procedure: Launch the asset "The call of the Wild"
press "Page_up" towards the end of asset, ensure it exits
back to linear.
Risks: Low

Signed-off-by: Kailash Baskar <kailash_baskar@comcast.com>
Change-Id: Ib51af7c61fcc520ead97ae75ae57e31e99adf24e