DELIA-43785 [AAMP][HLS][Westeros ON] The VOD playback freezes when trick play progressed in 2x rate
Reason for change: Issue reproducible only when there are multiple iframe profiles available.
During fragment timeout, profiles switched from lower to 2nd best and 2nd best to lower,
so the retry continuously happens for the failed fragment and not skipping the failed fragment in the process.
Now added a logic to avoid calling GetDesiredProfileBasedOnCache() or RampDownProfile()
if the lowest profile reached during rampdown, so the player will skip the failed fragment and try to pick the next fragment.
Test Procedure: Configured charles proxy and rewrite 200 response to curl 28 for any iframe fragment,
try to do trickplay before the iframe comes then see for the playback behavior, playback freezes for a moment then it will recover.
Risks: low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: Ibabfa84be0c2552f009d909f6f26db7b28464eec