DELIA-44074 CVR playback freezing,black screen and restarting
Reason for change: In muxed stream, discontinuity-EOS processing for audio track
ignored since it was tried to process before the first frame
and hence the mProcessingDiscontinuity(audio) is not set, but video discontinuity
for the same position processed and the flag was set. So the ProcessPendingDiscontinuity
ignored to process since both the tracks do not have discontinuity flag set and
due to stall on discontinuity processing schedule retune triggered.
Test Procedure: Tested by simulation.
Risks: Low
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>

Change-Id: I400cb18cc880cb9033ab152f4615c01e3422b3c8