DELIA-44124 [2006_Sprint] [AXG1V4] On launching Channel Store App "APPS-05114" error is observed.
Reason for change: Under ContentProtection tag, MarlinContentIds tag is considered as PSSH tag
in the ChannelStore content because of referring "string psshData = node.at(0)->GetText()" as pssh data,
so in this case MarlinContentIds is considered as PSSH and wrong DRM session was created
for ClearKey and that causes the licensing issue.
Added fix to check whether the tag is PSSH tag or not.
Test Procedure: As described in the ticket.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: If6a56a42b6fd8c4dec2b71e86357cb7b287c397f
(cherry picked from commit 3c2e4e07fc731742bcec2225d9aa0737b57cc260)