DELIA-45229 [sprint][DASH]: Crash with signature PrivateCDAIObjectMPD::ResetState
Reason for change: For DASH playback CDAIObjectMPD drived class object needs to be created,
but in this case CDAIObject base class object was created and that was not really having the mIsFogTSB member,
so when trying write on mIsFogTSB member with base class object under ResetState() it crashed.
Test Procedure: Play DASH content, it should not crash.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: Ib516633ee9672e51acee250b9307d32bd29038ca