DELIA-46198 Playback seeks to start position of TSB, after lightning FF
Reason for change: Issue observed when culled position reaches the paused position
in full TSB during the 4X speed and the player resumes the playback
since the condition checked for only for pipeline paused.
Fix checks if the player is in paused state with last known state = 4x trickplay (ie., lighting mode).
So, in this scenario, the player will not come out of Paused state, even if the culled position reaches paused position.
Test Procedure:
    Tune to IP Linear
    Wait for maximum TSB to build (25 minutes)
    Perform rewind to reach beginning of the TSB
    Again do fast forward in lightning speed to reach live position
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: I8522aa188bfb2e16269284a3bf36f20a876ba034

Change-Id: I537b53b434d0c9700c4374caefadb5f5cbd5cadc