DELIA-46385 : PEACOCK Playback exiting after an ad break.
Reason for change:  Player not able to pair the discontinuity between Video and Audio .
Some of the segments of content and Ad having duration > TargetDuration. Added a fix to handle
additional buffer duration +/- of 1sec to range check for discontinuity. Also added a warning log
for fragment duration > TargetDuration
Test Procedure: Playback of Peacock shows with Ads
Risks: Low

DEPS=46499

Signed-off-by: jkumar <Jai_Kumar@cable.comcast.com>
Change-Id: I75eb93824c607427ccf0816fb76b24c5a40eb4ed