DELIA-47453: sample live stream not playable with both <video> tag and UVE
Reason for change: Video playback got freezed due to PTS error, as
first and fresh Audio TS packet has been injected without PES data.
We handled this case by avoiding the PTS check of current and base
pts, when the current pts is not updated from the PES Packet.

Risks: Low

Signed-off-by: Supreeth Anil <supreeth.anil@ltts.com>
Change-Id: I6b0192ad69a4a46aa1c50f4007c602d3f5dd3751

DELIA-47453 - Switching to next audio track when current audio is not playing

Reason For Change: Audio is not playing in the particular hls file.
We always choose the first audio pid to play the audio content, even
if we have multiple aufio tracks.
When the first audfio pid is not having the valid PES data, then we
modified to choose the next audio pid and playing the audio data.

Risks: Low

Change-Id: I993f7c0b2fe0baf3592d35c0c0c483c9aa7d6a81
Signed-off-by: Madhavan Krishnan <madhavan.k@ltts.com>