DELIA-49304,DELIA-49367:Same set of timedMetadata pushed over playlist refresh
Reason for change: In this particular Peacock SLE, there are two metadata at
same timestamp. Due to this , every time new metadata comes a new entry is added as it fails to
match with timestamp and content.
Also included fix where full set of metadata was sent again on refresh when bulkMetadata is enabled
as two different tables were used to store .
Test Procedure: test with Peacock SLE
Risks: Medium

Signed-off-by: jkumar <Jai_Kumar@cable.comcast.com>
Change-Id: I3f9484e557dadfdc6567548c268d60fe11ae5582