DELIA-49763 [AAMP] [PEACOCK] Audio drop during Harmonic SLE
Reason for change: Discontinuity logic changed to wait for the next playlist refresh in case of no discontinuity in the other track when an audio/video track checks for paring.
Here the issue happened in a timing case where there is no discontinuity appended in audio track when video track has that, so video track was ignored as considered to be a singular discontinuity.
Fixed this issue.
Test Procedure: Playback harmonic SLE content and monitor for the audio drop
Risks: Low


Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: Ib4920818af239c9b257bcda9307394f8b2e8db02
(cherry picked from commit 82a04635ef5ce578fb9bcd2b1d24215440ab8338)