DELIA-49839 [AAMP] [PEACOCK] Init fragment missing in harmonic SLE
Reason for change: The issue was due to avoid updating the discontinuity flag for init due to a mismatch in the PDT of both the audio and video fragments.
so player logic checks if the other track discontinuity time is more than the current one, then it changes the playtarget to the bigger one and ignores updating the discontinuity, so the discontinuity flag is updated for the next fragment in the playlist, in this case, it's for the video fragment, so the init sent before the video got ignored to process.
and here is the original issue.
Test Procedure: Playback Harmonic MT SLE through Peacock ref app.
Risks: Low

Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: I6ecf4366f562b73e77e5d5b6d71d0ad14ef7156c
(cherry picked from commit 843f13d5d3ef54222290b9227ef675134ba536ab)