DELIA-48306 Pointer jump back and forth when doing trickplay
Reason for change: The issue was due to an older logic in the player to find the nearest resume position after trick play.
But now in the player, we have the latest logic which is based on GStreamer query positioning for the progress reporting.
Test Procedure: The progress pointer should not jump back and forth when doing trickplay in VOD and cDVR.
Risks: None

Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: Iac7d206f6d5074fd51abbe2290ebeb25cbb2358d