DELIA-51721 [FIELD][4.10] Peacock SLE events not starting to play
Reason for change: Since preferredLanguagesString is empty, there was a unnecessary teardown and tune initiated,
and during that time the because of saved PDT the previous PDT value got updated to 0
and hence the logic went wrong and calculated the invalid play target with negative value.
Because of the negative value, the playtarget modified to 0 that is why the SLE was not started with Live Edge.
Fixed both the language and playtarget issues.
Test Procedure: Playback Peacock SLE content and monitor whether the playback started at Live Point.
Risks: None
Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>



Change-Id: I7f61ff12fb8d56d35e4cf2fd7e03f41b0039f822
(cherry picked from commit 513fece7694f911f496e2c2e20d5b493573fe7ac)