DELIA-51654 [PureFlix] Browser crash when starting video
Reason for change: Due to the immediate stop after the seek call, Getfile() aborted and that returns error 42 and NULL fragment pointer, that NULL pointer passed to the ParseSegmentIndexBox() which is eventually crashing when tried to access it in Read32.
Test Procedure: Playback Pureflix VOD asset and try to do FF until to the end of the asset.
Risks: Low

Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
Change-Id: Ifd8e29ad1bec506bf3d1664353db5b1a50246691