DELIA-52372 : The get commands are giving mismatched results
Reason for change: The get command and set command are not in sync with the enum values.
Modified the get and set api to be insync with the enum values used for this.
Because of mismatch, the first eneum got ignored while showing the help and mismatched
with the numbers.
Test Procedure: Test the get and set command, and also added warning for set commands.
Risks: Low


Signed-off-by: Madhavan Krishnan <madhavan_Krishnan@comcast.com>
Change-Id: I1ae1b92c6f33c1d6947298e50ba203f68e955a60
(cherry picked from commit ae31400c75d93f308eadf4a040f7f28fc33a03f1)