DELIA-52007 Redbull Live video get freezed within 2 min of playback
Reason for change: Duration of Live is always returned as 1 min.
So webkit mediaplyer is pausing the playback when position goes
beyond the duration 1 min of playback in live case.
Now added culledseconds for calulating the duration for live.
Test Procedure: Test Live and VOD duration in Redbull App.
Risks: Low


Signed-off-by: Madhavan Krishnan <madhavan_Krishnan@comcast.com>
Change-Id: I407f42969dd36c8cdce4d20994a0a77367b16918