DELIA-55011 [AAMP] [PEACOCK] 30 tick due to discontinuity delta in Olympics Spotlight stream.
Reason for change: To mitigate the streamside ~4 sec discontinuity delta in the Player. Usually, 1 sec delta was allowed in the player to match the discontinuity pairing between audio and video; but with the current Olympic streams seeing ~4 sec delay because of that the discontinuity pairing was missed and the either of audio or video discontinuity was ignored.
Test Procedure: Playback Olympics Spotlight stream and monitor for "Comparing two disc date" print for diff value it should be more ore equal to one fragment duration which is more than 3 sec.
Risks: Low

Signed-off-by: Paulpandian M <paulpandian_mariappan@cable.comcast.com>
(cherry picked from commit 4cdbd3a5a53e4f02df4592bd9d317fe16c14a880)

Change-Id: I2874ffe7f23036754d4f17382e5e88d537b96f2c