DELIA-56341  :  [AAMP] [XUMO] [WIDEVINE] Free Movie crash
Reason for change: dash manifest does not content audio lang, hence represetation id is used as lang name
total len of lang string is 40 hence restricting it to 31 so that assert is not triggered when metadata event is sent.

Test Procedure: test XUMO ( which usaes aamp ) encypted content.

Risks: Low
Signed-off-by: shripad bankar <Shripad_Bankar@cable.comcast.com>
Change-Id: I37de63d58c85bffd2a686370422aed26422aabfc