DELIA-59549:aampcli 'auto' command crashes in ubuntu due to assertion while trying to open file for generated report
Reason for change: assert() when opening the test_results.csv file on UBUNTU
The reason is /opt is write protected on Ubuntu. If using sudo ./aamp-cli worked

Test Procedure: use auto command on Ubuntu. Without the fix it will stop at the first error that needs to be recorded

Risks: Medium

Signed-off-by: vesco vesselin_zlatev@comcast.com
Change-Id: I840b020212a55f64caf71d8226fc14d9fa1f000b